+++ ConsoleGame/Gui/Frame.cs
@@ -39,5 +39,4 @@ namespace ConsoleGame.Gui
             }
         }
     }
-
 }
