+++ ConsoleGame/View/GameWindow.cs
@@ -12,6 +12,7 @@ namespace ConsoleGame.View
         private Button creditsButton;
         private Button quitButton;
         private TextBlock titleTextBlock;
+        private List<Button> menuButtons = new List<Button>();
 
         public GameWindow() : base(0, 0, 120, 30, '%')
         {
@@ -24,7 +25,44 @@ namespace ConsoleGame.View
 
             quitButton = new Button(80, 13, 18, 5, "Quit");
 
+            menuButtons.Add(startButton);
+            menuButtons.Add(creditsButton);
+            menuButtons.Add(quitButton);
+            
             Render();
+
+            int hashCode;
+            do
+            {
+                hashCode = 39;
+                int i = 0;
+                while (hashCode == 39)
+                {
+                    if (i >= 0 && i < 3)
+                    {
+                        menuButtons[i].SetActive();
+                        Render();
+                        i++;
+                        menuButtons[i - 1].SetInActive();
+                    }
+                    ConsoleKeyInfo pressedChar = Console.ReadKey(true);
+                    hashCode = pressedChar.Key.GetHashCode();
+                }
+
+                while (hashCode == 37)
+                {
+                    if (i >= 0 && i < 3)
+                    {
+                        menuButtons[i].SetActive();
+                        Render();
+                        i--;
+                        menuButtons[i + 1].SetInActive();
+                    }
+                    ConsoleKeyInfo pressedChar = Console.ReadKey(true);
+                    hashCode = pressedChar.Key.GetHashCode();
+                }
+            } while (hashCode != 27);
+
         }
 
         public override void Render()
